executor: delete the repositories layer in favour of SeaORM directly
Every function was a one-line forward to the active-record API, which already
is the repository, and the boundary was fictional: the CLI imported
repositories directly and skipped services for nine of its calls.

update_status and update_config moved rather than inlined. update_status
carries double-Option semantics where an outer None means leave the column
alone, and update_config stamps updated_at while update_status must not,
because runs has no such column.
