+++ science-gateway/apps/executor/src/core/mod.rs
@@ -7,7 +7,6 @@ pub mod migrations;
 pub mod model_runners;
 pub mod output_locations;
 pub mod preflight;
-pub mod repositories;
 pub mod seed;
 pub mod services;
