--- science-gateway/apps/executor/src/core/repositories/mod.rs
@@ -1,8 +1,0 @@
-#![allow(dead_code)]
-
-pub mod artifact_types;
-pub mod artifacts;
-pub mod execution_targets;
-pub mod model_backends;
-pub mod model_invocation_profiles;
-pub mod runs;
