+++ science-gateway/apps/executor/src/core/services/artifact_types.rs
@@ -1,6 +1,8 @@
-use sea_orm::{DatabaseConnection, DbErr};
+use sea_orm::{
+    ActiveModelTrait, ColumnTrait, DatabaseConnection, DbErr, EntityTrait, QueryFilter, Set,
+};
 
-use crate::core::{entities::artifact_types, repositories};
+use crate::core::entities::artifact_types;
 
 use super::validation::require_json_object;
 
@@ -20,18 +22,32 @@ pub async fn register_artifact_type(
     input: RegisterArtifactTypeInput,
 ) -> Result<artifact_types::Model, DbErr> {
     require_json_object("artifact type metadata_schema", &input.metadata_schema_json)?;
-    repositories::artifact_types::create(db, input).await
+    artifact_types::ActiveModel {
+        slug: Set(input.slug),
+        label: Set(input.label),
+        default_format: Set(input.default_format),
+        display_mode: Set(input.display_mode),
+        viewer_kind: Set(input.viewer_kind),
+        description: Set(input.description),
+        metadata_schema_json: Set(input.metadata_schema_json),
+        ..Default::default()
+    }
+    .insert(db)
+    .await
 }
 
 pub async fn list_artifact_types(
     db: &DatabaseConnection,
 ) -> Result<Vec<artifact_types::Model>, DbErr> {
-    repositories::artifact_types::list(db).await
+    artifact_types::Entity::find().all(db).await
 }
 
 pub async fn get_artifact_type_by_slug(
     db: &DatabaseConnection,
     slug: &str,
 ) -> Result<Option<artifact_types::Model>, DbErr> {
-    repositories::artifact_types::find_by_slug(db, slug).await
+    artifact_types::Entity::find()
+        .filter(artifact_types::Column::Slug.eq(slug))
+        .one(db)
+        .await
 }
