+++ science-gateway/apps/executor/src/core/services/mod.rs
@@ -1,5 +1,3 @@
-#![allow(dead_code)]
-
 pub mod artifact_types;
 pub mod artifacts;
 pub mod execution_targets;
